refactor(parser): extract jest parser into shared package for vitest reuse
