+++ internal/javascript/parser/testdata/ancestors.json
