+++ internal/javascript/parser/testdata/empty.json
