+++ internal/javascript/parser/testdata/exec-error.json
