+++ internal/javascript/parser/testdata/fail.json
