+++ internal/javascript/parser/testdata/multi-file.json
