+++ internal/javascript/parser/testdata/null-duration.json
