+++ internal/javascript/parser/testdata/pass.json
