+++ internal/javascript/parser/testdata/pending.json
